+++ src/configurationsView.ts
@@ -23,11 +23,33 @@ export class ConfigViewProvider implements vscode.TreeDataProvider<ConfigOption>
 		{
 			for (let [key, values] of Object.entries(optionsList))
 			{
+				for (let i = 0; i < values.length; ++i)
+				{
+					if ((values[i] instanceof Object))
+					{
+						values[i] = new OptionPair(values[i]["name"], values[i]["value"]);
+					}
+					else
+					{
+						values[i] = new OptionPair(values[i], values[i]);
+					}
+				}
+				let currentValue;
 				if (!currentOptions.hasOwnProperty(key))
 				{
-					currentOptions[key] = values[0];
+					currentValue = values[0];
 				}
-				this.options.push(new ConfigOption(key, values, currentOptions[key]));
+				else
+				{
+					for (let i = 0; i < values.length; ++i)
+					{
+						if (values[i].value == currentOptions[key])
+						{
+							currentValue = values[i];
+						}
+					}
+				}
+				this.options.push(new ConfigOption(key, values, currentValue));
 			}
 		}
 		vscode.workspace.getConfiguration("launchOption").update("currentConfig", currentOptions, vscode.ConfigurationTarget.Workspace);
@@ -63,42 +85,49 @@ export class ConfigTreeOptions implements vscode.TreeViewOptions<ConfigOption> {
 
 const EmptyRow : string = "\"empty\"";
 
+class OptionPair {
+	constructor(public name : string, public value : string) {}
+}
 export class ConfigOption extends vscode.TreeItem {
+	tooltip : string;
+	description :string;
 
 	constructor(
 		public readonly label: string,
-		private values: string[],
-		private value: string
+		private values: OptionPair[],
+		private value : OptionPair
 	) {
-        super(label, vscode.TreeItemCollapsibleState.None);
+		super(label, vscode.TreeItemCollapsibleState.None);
+		this.value = value;
+		this.tooltip = `${this.label}-${this.value.name}`;
+		this.description = this.value.name == "" ? EmptyRow : this.value.name;
 	}
-
-	get tooltip(): string {
-		return `${this.label}-${this.value}`;
-	}
-
-	get description(): string {
-		return this.value == "" ? EmptyRow : this.value;
-    }
 
     public changeConfigParam(provider: ConfigViewProvider, context: vscode.ExtensionContext) {
-		let valuesToPick : Array<string> = this.values;
-		for (let i = 0; i < valuesToPick.length; ++i)
-			if (valuesToPick[i].length == 0)
-			valuesToPick[i] = EmptyRow
-        let newValue = vscode.window.showQuickPick(this.values);
+		let valuesToPick : Array<string> = Array<string>();
+		for (let i = 0; i < this.values.length; ++i)
+			valuesToPick.push(this.values[i].name.length == 0 ? EmptyRow : this.values[i].name);
+		let lineValue = this.value.name == "" ? EmptyRow : this.value.name;
+		valuesToPick.splice(valuesToPick.indexOf(lineValue), 1);
+		valuesToPick.unshift(lineValue);
+        let newValue = vscode.window.showQuickPick(valuesToPick);
         newValue.then((newVal)=>{
             if (newVal == undefined)
-                return;
-			this.value = newVal == EmptyRow ? "" : newVal;
+				return;
+			var foundValue = this.values.find((pair)=>{ return pair.name == (newVal == EmptyRow ? "" : newVal);});
+			if (foundValue == undefined)
+				return;
+			this.value = foundValue;
 			provider.refresh();
 			if (vscode.workspace.workspaceFolders?.length && this.label != undefined)
 			{
 				let config = vscode.workspace.getConfiguration("launchOption").get<LooseObject>("currentConfig");
 				if (config != undefined)
-					config[this.label] = this.value;
+					config[this.label] = this.value.value;
 				vscode.workspace.getConfiguration("launchOption").update("currentConfig", config, vscode.ConfigurationTarget.Workspace);
 			}
-        });
+			this.tooltip = `${this.label}-${this.value.name}`;
+			this.description = this.value.name == "" ? EmptyRow : this.value.name;
+		});
     }
 }